+++ js/main.js
@@ -1,56 +1,5 @@
-const MESSAGES = [
-  'Всё отлично!',
-  'В целом всё неплохо. Но не всё.',
-  'Когда вы делаете фотографию, хорошо бы убирать палец из кадра.',
-  'Моя бабушка случайно чихнула и у неё получилась фотография лучше.',
-  'Я уронил фотоаппарат на кота и получилось лучше.',
-  'Лица перекошены, как будто их избивают. Как можно было поймать такой момент?!',
-];
+import { generatePhotos } from './data.js';
 
-const NAMES = [
-  'Иван',
-  'Мария', 
-  'Сергей',
-  'Юлия', 
-  'Олег',
-  'Анна', 
-  'Максим', 
-  'Алексей',
-  'Ольга', 
-  'Виктор',
-];
-
-const getRandomInteger = (a, b) => {
-  const lower = Math.ceil(Math.min(a, b));
-  const upper = Math.floor(Math.max(a, b));
-  const result = Math.random() * (upper - lower + 1) + lower;
-  return Math.floor(result);
-};
-
-const getRandomArrayElement = (array) => array[getRandomInteger(0, array.length - 1)];
-
-let commentId = 1;
-
-const createComment = () => ({
-  id: commentId++,
-  avatar: `img/avatar-${getRandomInteger(1, 6)}.svg`,
-  message: Array.from(
-    { length: getRandomInteger(1, 2) },
-    () => getRandomArrayElement(MESSAGES)
-  ).join(' '),
-  name: getRandomArrayElement(NAMES),
-});
-
-const createPhoto = (id) => ({
-  id: id,
-  url: `photos/${id}.jpg`,
-  description: `Это фотография №${id}`,
-  likes: getRandomInteger(15, 200),
-  comments: Array.from({ length: getRandomInteger(0, 30) }, createComment),
-});
-
-const PHOTOS_COUNT = 25;
-
-const photos = Array.from({ length: PHOTOS_COUNT }, (_, index) => createPhoto(index + 1));
+const photos = generatePhotos();
 
 console.log(photos);
